Wave Spawning
Wave spawning is managed by GameManager.cs, and functions by instantiating gameobjects randomly across however many spawners exist in the game. If none exist, but a spawn is expected, it will throw an error.

The number of enemies on the screen are also tracked using an increase of value upon instantation, and a decrease at Death();. The effected value is "enemies" and is located in PointCounter.cs